+++ GUI/MapEditor.cpp
@@ -6,9 +6,9 @@
 #include <FL/Fl_Return_Button.H>
 #include "MapEditor.h"
 #include "MapSerializer.h"
-#include "../ProjectSourceCode/Interactable.h"
-#include "../ProjectSourceCode/Wall.h"
-#include "../ProjectSourceCode/EmptyCell.h"
+#include "../ProjectSourceCode/Interactable/Interactable.h"
+#include "../ProjectSourceCode/Interactable/Wall.h"
+#include "../ProjectSourceCode/Interactable/EmptyCell.h"
 #include "../ProjectSourceCode/Character.h"
 #include "../ProjectSourceCode/item.h"
 #include "../ProjectSourceCode/MapBuilder.h"
