+++ [COMP345] Assignment 1.vcxproj
@@ -124,10 +124,10 @@
     <ClCompile Include="ProjectSourceCode\Strategy\aggressorstrategy.cpp" />
     <ClCompile Include="ProjectSourceCode\Strategy\friendlystrategy.cpp" />
     <!-- <ClCompile Include="ProjectSourceCode\testSerializeCharacter.cpp" /> -->
-    <ClCompile Include="ProjectSourceCode\EmptyCell.cpp" />
-    <ClCompile Include="ProjectSourceCode\Interactable.cpp" />
+    <ClCompile Include="ProjectSourceCode\Interactable\EmptyCell.cpp" />
+    <ClCompile Include="ProjectSourceCode\Interactable\Interactable.cpp" />
     <ClCompile Include="ProjectSourceCode\MapBuilder.cpp" />
-    <ClCompile Include="ProjectSourceCode\Wall.cpp" />
+    <ClCompile Include="ProjectSourceCode\Interactable\Wall.cpp" />
     <ClCompile Include="ProjectSourceCode\Game\game.cpp" />
     <ClCompile Include="ProjectSourceCode\Game\gamelogger.cpp" />
     <ClCompile Include="ProjectSourceCode\testgamelogger.cpp" />
@@ -171,10 +171,10 @@
     <ClInclude Include="ProjectSourceCode\Strategy\humanplayerstrategy.h" />
     <ClInclude Include="ProjectSourceCode\Strategy\aggressorstrategy.h" />
     <ClInclude Include="ProjectSourceCode\Strategy\friendlystrategy.h" />
-    <ClInclude Include="ProjectSourceCode\EmptyCell.h" />
-    <ClInclude Include="ProjectSourceCode\Interactable.h" />
+    <ClInclude Include="ProjectSourceCode\Interactable\EmptyCell.h" />
+    <ClInclude Include="ProjectSourceCode\Interactable\Interactable.h" />
     <ClInclude Include="ProjectSourceCode\MapBuilder.h" />
-    <ClInclude Include="ProjectSourceCode\Wall.h" />
+    <ClInclude Include="ProjectSourceCode\Interactable\Wall.h" />
     <ClInclude Include="ProjectSourceCode\Game\game.h" />
     <ClInclude Include="ProjectSourceCode\Game\gamelogger.h" />
     <ClInclude Include="ProjectSourceCode\testgamelogger.h" />
