+++ [COMP345] Assignment 1.vcxproj.filters
@@ -81,16 +81,16 @@
     <ClCompile Include="ProjectSourceCode\Strategy\friendlystrategy.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ProjectSourceCode\EmptyCell.cpp">
+    <ClCompile Include="ProjectSourceCode\Interactable\EmptyCell.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ProjectSourceCode\Interactable.cpp">
+    <ClCompile Include="ProjectSourceCode\Interactable\Interactable.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="ProjectSourceCode\MapBuilder.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ProjectSourceCode\Wall.cpp">
+    <ClCompile Include="ProjectSourceCode\Interactable\Wall.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="ProjectSourceCode\Game\game.cpp">
@@ -203,16 +203,16 @@
     <ClInclude Include="ProjectSourceCode\Strategy\friendlystrategy.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ProjectSourceCode\EmptyCell.h">
+    <ClInclude Include="ProjectSourceCode\Interactable\EmptyCell.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ProjectSourceCode\Interactable.h">
+    <ClInclude Include="ProjectSourceCode\Interactable\Interactable.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ProjectSourceCode\MapBuilder.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ProjectSourceCode\Wall.h">
+    <ClInclude Include="ProjectSourceCode\Interactable\Wall.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="GUI\BaseEditor.h">
